+++ actix-http/src/body/mod.rs
@@ -11,7 +11,6 @@ use futures_core::ready;
 mod body;
 mod body_stream;
 mod message_body;
-mod response_body;
 mod size;
 mod sized_stream;
 
@@ -19,7 +18,6 @@ pub use self::body::{AnyBody, Body, BoxBody};
 pub use self::body_stream::BodyStream;
 pub use self::message_body::MessageBody;
 pub(crate) use self::message_body::MessageBodyMapErr;
-pub use self::response_body::ResponseBody;
 pub use self::size::BodySize;
 pub use self::sized_stream::SizedStream;