+++ src/responder.rs
@@ -232,7 +232,7 @@ pub(crate) mod tests {
     use bytes::{Bytes, BytesMut};
 
     use super::*;
-    use crate::dev::{Body, ResponseBody};
+    use crate::dev::AnyBody;
     use crate::http::{header::CONTENT_TYPE, HeaderValue, StatusCode};
     use crate::test::{init_service, TestRequest};
     use crate::{error, web, App};
@@ -264,39 +264,18 @@ pub(crate) mod tests {
 
     pub(crate) trait BodyTest {
         fn bin_ref(&self) -> &[u8];
-        fn body(&self) -> &Body;
+        fn body(&self) -> &AnyBody;
     }
 
     impl BodyTest for Body {
         fn bin_ref(&self) -> &[u8] {
             match self {
-                Body::Bytes(ref bin) => bin,
+                AnyBody::Bytes(ref bin) => bin,
                 _ => unreachable!("bug in test impl"),
             }
         }
         fn body(&self) -> &Body {
             self
-        }
-    }
-
-    impl BodyTest for ResponseBody<Body> {
-        fn bin_ref(&self) -> &[u8] {
-            match self {
-                ResponseBody::Body(ref b) => match b {
-                    Body::Bytes(ref bin) => bin,
-                    _ => unreachable!("bug in test impl"),
-                },
-                ResponseBody::Other(ref b) => match b {
-                    Body::Bytes(ref bin) => bin,
-                    _ => unreachable!("bug in test impl"),
-                },
-            }
-        }
-        fn body(&self) -> &Body {
-            match self {
-                ResponseBody::Body(ref b) => b,
-                ResponseBody::Other(ref b) => b,
-            }
         }
     }